updated dependenciesto angular 8 boilerplate updated according newer dotnet angular template
