--- Frontend/ClientApp/e2e/app.e2e-spec.ts
@@ -1,14 +1,0 @@
-import { AppPage } from './app.po';
-
-describe('App', () => {
-  let page: AppPage;
-
-  beforeEach(() => {
-    page = new AppPage();
-  });
-
-  it('should display welcome message', () => {
-    page.navigateTo();
-    expect(page.getMainHeading()).toEqual('Hello, world!');
-  });
-});
