--- Frontend/ClientApp/e2e/app.po.ts
@@ -1,11 +1,0 @@
-import { browser, by, element } from 'protractor';
-
-export class AppPage {
-  navigateTo() {
-    return browser.get('/');
-  }
-
-  getMainHeading() {
-    return element(by.css('app-root h1')).getText();
-  }
-}
