+++ Frontend/ClientApp/package.json
@@ -6,59 +6,55 @@
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build",
-    "build:ssr": "npm run build -- --app=ssr --output-hashing=media",
+    "build:ssr": "ng run dotnet_angular:server:dev",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "7.2.9",
-    "@angular/common": "7.2.9",
-    "@angular/compiler": "7.2.9",
-    "@angular/core": "7.2.9",
-    "@angular/forms": "7.2.9",
-    "@angular/http": "7.2.9",
-    "@angular/platform-browser": "7.2.9",
-    "@angular/platform-browser-dynamic": "7.2.9",
-    "@angular/platform-server": "7.2.9",
-    "@angular/router": "7.2.9",
-    "@ng-bootstrap/ng-bootstrap": "^4.1.0",
-    "@nguniversal/module-map-ngfactory-loader": "^7.1.1",
+    "@angular/animations": "~8.1.3",
+    "@angular/common": "~8.1.3",
+    "@angular/compiler": "~8.1.3",
+    "@angular/core": "~8.1.3",
+    "@angular/forms": "~8.1.3",
+    "@angular/platform-browser": "~8.1.3",
+    "@angular/platform-browser-dynamic": "~8.1.3",
+    "@angular/platform-server": "~8.1.3",
+    "@angular/router": "~8.1.3",
+    "@ng-bootstrap/ng-bootstrap": "5.1.0",
+    "@nguniversal/module-map-ngfactory-loader": "8.1.1",
     "@ngx-loading-bar/core": "^4.2.0",
     "aspnet-prerendering": "^3.0.1",
     "bootstrap": "^4.3.1",
-    "chartist": "^0.11.0",
-    "core-js": "^2.6.5",
-    "ol": "^5.3.1",
-    "rxjs": "^6.4.0",
-    "tslib": "^1.9.3",
-    "zone.js": "~0.8.29"
+    "chartist": "^0.11.3",
+    "core-js": "^3.1.4",
+    "ol": "^5.3.3",
+    "rxjs": "~6.5.2",
+    "tslib": "^1.10.0",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.5",
-    "@angular/cli": "^7.3.5",
-    "@angular/compiler-cli": "7.2.9",
-    "@angular/language-service": "7.2.9",
-    "@types/chartist": "^0.9.44",
-    "@types/jasmine": "^3.3.9",
-    "@types/jasminewd2": "^2.0.6",
-    "@types/node": "^11.11.3",
-    "codelyzer": "^4.5.0",
-    "jasmine-core": "~3.3.0",
+    "@angular-devkit/build-angular": "~0.801.2",
+    "@angular/cli": "~8.1.2",
+    "@angular/compiler-cli": "~8.1.3",
+    "@angular/language-service": "8.1.3",
+    "@types/chartist": "~0.9.46",
+    "@types/jasmine": "~3.3.16",
+    "@types/jasminewd2": "~2.0.6",
+    "@types/node": "~12.6.8",
+    "codelyzer": "^5.1.0",
+    "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^4.0.1",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^2.0.5",
+    "karma": "~4.2.0",
+    "karma-chrome-launcher": "~3.0.0",
+    "karma-coverage-istanbul-reporter": "~2.1.0",
     "karma-jasmine": "^2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
+    "karma-jasmine-html-reporter": "^1.4.2",
     "protractor": "~5.4.2",
-    "ts-node": "~8.0.3",
-    "tslint": "~5.14.0",
-    "typescript": "3.2.4",
-    "typescript-tslint-plugin": "^0.3.1"
-  },
-  "optionalDependencies": {
-    "node-sass": "^4.9.3"
+    "ts-node": "~7.0.0",
+    "tslint": "~5.18.0",
+    "typescript": "~3.4.5",
+    "typescript-tslint-plugin": "^0.5.4"
   }
 }
