+++ Frontend/Startup.cs
@@ -45,6 +45,7 @@ namespace Frontend
             else
             {
                 app.UseExceptionHandler("/Error");
+                // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
                 app.UseHsts();
             }
